+++ spec/support/shared_examples/__axn_mounted_to__behavior.rb
@@ -43,7 +43,7 @@ RSpec.shared_examples "__axn_mounted_to__ behavior" do |attachment_method|
     end
 
     before do
-      if attachment_method == :axn_method
+      if attachment_method == :mount_axn_method
         client_class.public_send(attachment_method, :existing_action, axn_klass: existing_axn)
       else
         client_class.public_send(attachment_method, :existing_action, existing_axn)