+++ spec_rails/dummy_app/app/actions/clients/user.rb
@@ -4,7 +4,7 @@ module Actions::Clients
   class User
     include Axn
 
-    axn_method(:get_name) { |id:| user(id:).name }
+    mount_axn_method(:get_name) { |id:| user(id:).name }
     axn(:email, expose_return_as: :value) { |id:| user(id:).email }
 
     private